+++ src/main/java/es/jbp/kajtools/templates/symbols/RandFunction.java
@@ -1,6 +1,7 @@
 package es.jbp.kajtools.templates.symbols;
 
 import es.jbp.expressions.Value;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -10,7 +11,7 @@ public class RandFunction extends AbstractFunction {
   public Value evaluate(List<Value> parameterList) {
     long limit = getParameterAsLong(parameterList, 0, 10L);
     long result = ThreadLocalRandom.current().nextLong(0, limit);
-    return new Value(result);
+    return new Value(BigDecimal.valueOf(result));
   }
 
   @Override
